# openapi.yaml
openapi: "3.1.0"
info:
  title: SHL Data API
  description: |
    Read-only API for SHL (Swedish Hockey League) data scraped from SweHockey.
    Provides schedule, standings, game details, player/goalie stats, and team rosters.
  version: "1.0.0"
  contact:
    name: shl2026-backend
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /seasons/{season_id}/schedule:
    get:
      operationId: getSchedule
      summary: Full season schedule
      parameters:
        - $ref: "#/components/parameters/SeasonId"
      responses:
        "200":
          description: Season schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /seasons/{season_id}/games:
    get:
      operationId: getGamesByDate
      summary: Games filtered by date
      parameters:
        - $ref: "#/components/parameters/SeasonId"
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2025-09-13"
      responses:
        "200":
          description: Games for the specified date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleResponse"
  /seasons/{season_id}/games/today:
    get:
      operationId: getTodaysGames
      summary: Today's upcoming/live games
      parameters:
        - $ref: "#/components/parameters/SeasonId"
      responses:
        "200":
          description: Today's games without a final result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleResponse"
...
  /seasons/{season_id}/games/played:
    get:
      operationId: getPlayedGames
      summary: All played games (with results)
      parameters:
        - $ref: "#/components/parameters/SeasonId"
      responses:
        "200":
          description: All games with a final result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleResponse"
  /seasons/{season_id}/rounds:
    get:
      operationId: getRounds
      summary: Schedule grouped by round
      parameters:
        - $ref: "#/components/parameters/SeasonId"
      responses:
        "200":
          description: All rounds with their games
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundsResponse"
  /seasons/{season_id}/rounds/played:
    get:
      operationId: getPlayedRounds
      summary: Completed rounds only
      parameters:
        - $ref: "#/components/parameters/SeasonId"
      responses:
        "200":
          description: Rounds where all games have results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundsResponse"
  /seasons/{season_id}/rounds/next:
    get:
      operationId: getNextRound
      summary: Next unplayed round
      parameters:
        - $ref: "#/components/parameters/SeasonId"
      responses:
        "200":
          description: Next round with unplayed games
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /seasons/{season_id}/standings:
    get:
      operationId: getStandings
      summary: League standings (live during games)
      description: |
        Computed from schedule data. During game days, standings reflect
        in-progress scores within ~30 seconds of a goal.
        Pre-season (no games played): returns all teams at rank 1, sorted alphabetically.
      parameters:
        - $ref: "#/components/parameters/SeasonId"
      responses:
        "200":
          description: Current standings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandingsResponse"
  /seasons/{season_id}/players:
    get:
      operationId: getPlayerStats
      summary: Scoring leaders
      parameters:
        - $ref: "#/components/parameters/SeasonId"
        - name: team
          in: query
          required: false
          description: Filter by team abbreviation (e.g. SKE, FRÖ)
          schema:
            type: string
      responses:
        "200":
          description: Player scoring statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerStatsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /seasons/{season_id}/goalies:
    get:
      operationId: getGoalieStats
      summary: Goalie statistics
      parameters:
        - $ref: "#/components/parameters/SeasonId"
        - name: team
          in: query
          required: false
          description: Filter by team abbreviation
          schema:
            type: string
      responses:
        "200":
          description: Goalie statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalieStatsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /seasons/{season_id}/teams:
    get:
      summary: List teams with abbreviations for a season
      operationId: getSeasonTeams
      parameters:
        - name: season_id
          in: path
          required: true
          schema:
            type: integer
          description: SweHockey season/tournament ID
      responses:
        '200':
          description: Team list with abbreviations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TeamInfo'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '404':
          description: Team info not yet fetched
  /seasons/{season_id}/players/{team}/{jersey}/stats:
    get:
      operationId: getPlayerStats
      summary: Dynamic game stats for a single player
      description: |
        Returns detailed game statistics for a player from the SweHockey PlayersByTeam page.
        Data is fetched on demand (not cached long-term) since stats change frequently.
        Includes GP, G, A, TP, PIM, +/-, GWG, PPG, SHG, SOG, SG%, faceoffs.
      parameters:
        - $ref: "#/components/parameters/SeasonId"
        - name: team
          in: path
          required: true
          schema:
            type: string
          description: SweHockey team name (e.g. "Brynäs IF")
          example: "Brynäs IF"
        - name: jersey
          in: path
          required: true
          schema:
            type: integer
          description: Player jersey number
          example: 3
      responses:
        "200":
          description: Player game statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerStatsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Failed to fetch stats from upstream
  /seasons/{season_id}/players/{team}/{jersey}:
    get:
      operationId: getMergedPlayerDetail
      summary: Single player with merged SweHockey roster + shl.se data
      description: |
        Returns a combined view of a player, merging SweHockey roster data
        (birthdate, height, weight, handedness, youth club) with shl.se enrichment
        (portrait URL, team code, shl.se UUID). Both sources are lazy-loaded on demand.
      parameters:
        - $ref: "#/components/parameters/SeasonId"
        - name: team
          in: path
          required: true
          schema:
            type: string
          description: SweHockey team name (e.g. "Brynäs IF")
          example: "Brynäs IF"
        - name: jersey
          in: path
          required: true
          schema:
            type: integer
          description: Player jersey number
          example: 3
      responses:
        "200":
          description: Merged player detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MergedPlayerResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Failed to fetch data from upstream
  /seasons/{season_id}/players/{team}:
    get:
      operationId: getMergedTeamPlayers
      summary: All players for a team with merged SweHockey + shl.se data
      description: |
        Returns all players on a team roster, merging SweHockey roster data with
        shl.se enrichment (portraits, UUIDs). Both sources are lazy-loaded on demand.
      parameters:
        - $ref: "#/components/parameters/SeasonId"
        - name: team
          in: path
          required: true
          schema:
            type: string
          description: SweHockey team name (e.g. "Brynäs IF")
          example: "Brynäs IF"
      responses:
        "200":
          description: Merged team player list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MergedTeamPlayersResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Failed to fetch roster data from upstream
...
  /games/{game_id}:
    get:
      operationId: getGameDetail
      summary: Game detail (score, events, team stats)
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full game detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /devices:
    post:
      operationId: registerDevice
      summary: Register device for push notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRegistration"
      responses:
        "200":
          description: Device registered
    delete:
      operationId: unregisterDevice
      summary: Unregister device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceUnregistration"
      responses:
        "200":
          description: Device unregistered
components:
  parameters:
    SeasonId:
      name: season_id
      in: path
      required: true
      schema:
        type: integer
        example: 18263
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
    Meta:
      type: object
      description: Response metadata with freshness info
      properties:
        generated_at:
          type: string
          format: date-time
        season_id:
          type: string
        source_fetched_at:
          type: string
          format: date-time
...
    TeamInfo:
      type: object
      properties:
        team:
          type: string
          description: Full team name
          example: "Brynäs IF"
        abbreviation:
          type: string
          description: Official team abbreviation
          example: "BIF"
      required:
        - team
        - abbreviation
...
    DeviceRegistration:
      type: object
      required:
        - fcm_token
      properties:
        fcm_token:
          type: string
        platform:
          type: string
          default: android
...
    ScheduleResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleEntry"
        meta:
          $ref: "#/components/schemas/Meta"
    StandingsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/StandingsRow"
        meta:
          $ref: "#/components/schemas/Meta"
...
    GoalieStatsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/GoalieStat"
        meta:
          $ref: "#/components/schemas/Meta"
    RoundsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Round"
        meta:
          $ref: "#/components/schemas/Meta"
    RoundResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Round"
        meta:
          $ref: "#/components/schemas/Meta"
    GameDetailResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            game:
              $ref: "#/components/schemas/GameInfo"
            score:
              $ref: "#/components/schemas/Score"
            teams:
              type: object
              additionalProperties:
                type: object
                description: Team stats (shots, saves, pim, pp)
            actions:
              type: array
              items:
                $ref: "#/components/schemas/Action"
        meta:
          $ref: "#/components/schemas/Meta"
...
    MergedPlayerResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/MergedPlayer"
        meta:
          $ref: "#/components/schemas/Meta"
    MergedTeamPlayersResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MergedPlayer"
        meta:
          allOf:
            - $ref: "#/components/schemas/Meta"
            - type: object
              properties:
                count:
                  type: integer
                  description: Number of players returned
...
    PlayerStatsResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/TeamPlayerStat"
        meta:
          $ref: "#/components/schemas/Meta"